% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qSOFA.r
\name{qSOFA}
\alias{qSOFA}
\title{Calculate qSOFA(quick Sepsis-related Organ Failure Assessment) following SEPSIS GUIDELINE 3
You need input 3 values
GCS =< 13 : input x = 1
RR >= 2 : input y = 1
sBP <= 100 : input z = 1}
\usage{
qSOFA(x, y, z)
}
\arguments{
\item{caluation}{qSOFA}
}
\description{
Calculate qSOFA(quick Sepsis-related Organ Failure Assessment) following SEPSIS GUIDELINE 3
You need input 3 values
GCS =< 13 : input x = 1
RR >= 2 : input y = 1
sBP <= 100 : input z = 1
}
\examples{
qSOFA ()
}
\keyword{qSOFA}
